fix(library/tactic/simplify): avoid an expensive copy in simp (#801)
In that context, `auto` is deduced to be `expr_map<simp_result>`, which creates a local copy of the map. However, the call to `find()` below can be made on a `const&`, so the copy is unnecessary.

We've measured this copy to be using ~1% of total runtime.

This is intended as a non-functional change. It was generated by automated tools.